Avg of first path

+++ src/main/java/org/soffredi/api/AvgAndDev.java
@@ -3,7 +3,7 @@ package org.soffredi.api;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class AvgAndDev {
-    private int avg;
+    private double avg;
     private double dev;
 
     public AvgAndDev() {
@@ -11,13 +11,13 @@ public class AvgAndDev {
         dev = 0;
     }
 
-    public AvgAndDev(int avg, double dev) {
+    public AvgAndDev(double avg, double dev) {
         this.avg = avg;
         this.dev = dev;
     }
 
     @JsonProperty
-    public int getAvg() {
+    public double getAvg() {
         return avg;
     }
